+++ conf/config.js
@@ -1,19 +1,7 @@
 module.exports = {
-	titanium: {
-		sdk: {
-			/**
-			 * The default Titanium SDK installation path.
-			 * @type {String}
-			 */
-			defaultInstallLocation: null
-		},
-
-		modules: {
-			/**
-			 * The default Titanium Modules installation path.
-			 * @type {String}
-			 */
-			defaultInstallLocation: null
-		}
-	}
+	/**
+	 * The default Titanium SDK installation path.
+	 * @type {Array.<String>}
+	 */
+	searchPaths: []
 };