+++ src/cli/commands/sdk.js
@@ -1,0 +1,6 @@
+export default {
+	action({ console }) {
+		console.log('SDKs!');
+	},
+	desc: 'Manage Titanium SDKs'
+};