+++ src/module/module-list-service.js
@@ -2,7 +2,7 @@ import DetectEngine from 'appcd-detect';
 import gawk from 'gawk';
 import sortObject from 'sort-object-keys';
 
-import { compare } from '../version';
+import { compare } from '../lib/version';
 import { DataServiceDispatcher } from 'appcd-dispatcher';
 import { debounce, get } from 'appcd-util';
 import { modules, options } from 'titaniumlib';